# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.PostAuthorDTO:
    properties:
      admin:
        type: boolean
      avatar:
        type: string
      id:
        type: integer
      nickname:
        type: string
    type: object
  dto.PostCreateInput:
    properties:
      content:
        type: string
      disable_comment:
        type: boolean
      hide:
        type: boolean
      pin:
        type: integer
    type: object
  dto.PostDTO:
    properties:
      author:
        $ref: '#/definitions/dto.PostAuthorDTO'
      content:
        type: string
      created_at:
        type: string
      disable_comment:
        type: boolean
      hide:
        type: boolean
      id:
        type: integer
      pin:
        type: integer
      tags:
        items:
          $ref: '#/definitions/dto.TagDTO'
        type: array
      updated_at:
        type: string
    type: object
  dto.PostListResult:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.PostDTO'
        type: array
      total:
        type: integer
    type: object
...
  dto.PostUpdateInput:
    properties:
      clear_pin:
        type: boolean
      content:
        type: string
      disable_comment:
        type: boolean
      hide:
        type: boolean
      pin:
        type: integer
    type: object
...
  dto.SettingItem:
    properties:
      category:
        type: string
      desc:
        type: string
      key:
        type: string
      public:
        type: boolean
      sensitive:
        type: boolean
      value:
        type: string
    type: object
  dto.SettingListResult:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.SettingItem'
        type: array
      total:
        type: integer
    type: object
  dto.SettingPublicItem:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
  dto.SettingPublicResult:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.SettingPublicItem'
        type: array
    type: object
  dto.SettingUpdateInput:
    properties:
      value:
        type: string
    type: object
  dto.TagDTO:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 这是一个轻量级的说说服务
  title: EasyDrop API
  version: "1.0"
paths:
...
  /admin/posts:
    get:
      description: 分页查询说说列表，支持按用户和标签过滤
      parameters:
      - description: 用户ID
        in: query
        name: user_id
        type: integer
      - description: 标签ID
        in: query
        name: tag_id
        type: integer
      - description: 内容关键字
        in: query
        name: content
        type: string
      - description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      - description: 每页条数，最大 100
        in: query
        name: size
        type: integer
      - description: 排序，如 created_at desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PostListResult'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 管理端查询说说列表
      tags:
      - post-admin
    post:
      consumes:
      - application/json
      description: 创建说说并返回详情
      parameters:
      - description: 说说创建参数
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.PostCreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PostDTO'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 管理端创建说说
      tags:
      - post-admin
  /admin/posts/{id}:
    delete:
      description: 根据说说 ID 删除说说
      parameters:
      - description: 说说ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: 说说不存在
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 管理端删除说说
      tags:
      - post-admin
    get:
      description: 根据说说 ID 查询详情
      parameters:
      - description: 说说ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PostDTO'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: 说说不存在
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 管理端查询说说详情
      tags:
      - post-admin
    patch:
      consumes:
      - application/json
      description: 根据说说 ID 更新说说内容
      parameters:
      - description: 说说ID
        in: path
        name: id
        required: true
        type: integer
      - description: 说说更新参数
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.PostUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PostDTO'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: 说说不存在
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 管理端更新说说
      tags:
      - post-admin
  /admin/settings:
    get:
      description: 分页查询配置列表，支持按分类和键名过滤
      parameters:
      - description: 分类
        in: query
        name: category
        type: string
      - description: 键名（模糊匹配）
        in: query
        name: key
        type: string
      - description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      - description: 每页条数，最大 100
        in: query
        name: size
        type: integer
      - description: 排序，如 key_asc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SettingListResult'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 管理端查询配置列表
      tags:
      - setting-admin
  /admin/settings/{key}:
    patch:
      consumes:
      - application/json
      description: 按配置 key 更新 value
      parameters:
      - description: 配置键
        in: path
        name: key
        required: true
        type: string
      - description: 更新参数
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.SettingUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 管理端更新配置
      tags:
      - setting-admin
...
  /settings/public:
    get:
      description: 返回前端可公开读取的站点配置
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SettingPublicResult'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 前端公开配置
      tags:
      - setting
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"